Cleared up some bugs with Image Window.
git-svn-id: file:///Users/wmohler/Documents/CytoSHOWsvnClone/svn_load_compat@18583 d59c7341-666b-7e48-8bb3-2014e8cfb900

+++ GLOWormJ022614/GLOWormJviaperlSVN/GLOWormJviaperlSVN32last/src/ij/gui/ImageCanvas.java
@@ -124,7 +124,7 @@ public class ImageCanvas extends Canvas implements MouseListener, MouseMotionLis
 		imageWidth = width;
 		imageHeight = height;
 		srcRect = new Rectangle(0, 0, imageWidth, imageHeight);
-		setDrawingSize(imageWidth, (int)(imageHeight));
+		setDrawingSize(imageWidth, imageHeight);
 		magnification = 1.0;
 		messageRois = new Hashtable<String,Roi>();
